Issue #4590 fixed

+++ packages/Webkul/Shop/src/Http/Controllers/OnepageController.php
@@ -74,7 +74,7 @@ class OnepageController extends Controller
             return redirect()->route('customer.session.index');
         }
 
-        $minimumOrderAmount = core()->getConfigData('sales.orderSettings.minimum-order.minimum_order_amount') ?? 0;
+        $minimumOrderAmount = (float) core()->getConfigData('sales.orderSettings.minimum-order.minimum_order_amount') ?? 0;
 
         if (! $cart->checkMinimumOrder()) {
             session()->flash('warning', trans('shop::app.checkout.cart.minimum-order-message', ['amount' => core()->currency($minimumOrderAmount)]));
